+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `grafana-organization-reconciler`.
+
 ## [0.22.0] - 2025-02-12
 
 ### Changed